git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@1893 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/sidtune/SidTuneBase.cpp
@@ -457,9 +457,11 @@ void SidTuneBase::convertOldStyleSpeedToTables(uint_least32_t speed, SidTuneInfo
 {
     // Create the speed/clock setting tables.
     //
-    // This routine implements the PSIDv2NG compliant speed conversion.  All tunes
+    // This routine implements the PSIDv2NG compliant speed conversion. All tunes
     // above 32 use the same song speed as tune 32
-    const unsigned int toDo = std::min(info->m_songs, MAX_SONGS);
+    // NOTE: The cast here is used to avoid undefined references
+    // as the std::min function takes it parameters by reference
+    const unsigned int toDo = std::min(info->m_songs, static_cast<unsigned int>(MAX_SONGS));
     for (unsigned int s = 0; s < toDo; s++)
     {
         clockSpeed[s] = clock;
